Finished data layer and implemented diffUtils

+++ core/src/main/java/xyz/savvamirzoyan/nous/core/TypeAlias.kt
@@ -1,4 +1,4 @@
 package xyz.savvamirzoyan.nous.core
 
-typealias ID = Int
+typealias ID = Long
 typealias PictureUrl = String
